Add SSE keep-alive pings for universal proxy compatibility
Extract shared SSE setup into sseHelper. Send an SSE comment (: ping)
every 5 seconds to keep connections alive through any reverse proxy
(nginx, Apache, Traefik, Synology Web Station etc.) and force buffer
flushes even on proxies that ignore X-Accel-Buffering.
