+++ src/pages/api/import-splitpro-stream.ts
@@ -4,6 +4,7 @@ import { getServerSession } from 'next-auth/next';
 import { type File, formidable } from 'formidable';
 import { authOptions } from '~/server/auth';
 import { importSplitProData, restoreSplitProData } from '~/server/api/services/splitService';
+import { setupSSE } from '~/server/sseHelper';
 
 export const config = {
   api: { bodyParser: false, responseLimit: false },
@@ -19,23 +20,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
     return res.status(401).json({ error: 'Unauthorized' });
   }
 
-  res.setHeader('Content-Type', 'text/event-stream');
-  res.setHeader('Cache-Control', 'no-cache, no-transform');
-  res.setHeader('Connection', 'keep-alive');
-  res.setHeader('X-Accel-Buffering', 'no');
-  res.setHeader('Content-Encoding', 'none');
-  res.flushHeaders();
-
-  const flush = () => {
-    if ('flush' in res && typeof res.flush === 'function') {
-      (res as unknown as { flush: () => void }).flush();
-    }
-  };
-
-  const send = (type: string, payload: object) => {
-    res.write(`data: ${JSON.stringify({ type, ...payload })}\n\n`);
-    flush();
-  };
+  const { send, end } = setupSSE(res);
 
   let uploadedFile: File | undefined;
   try {
@@ -46,7 +31,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
     if (!uploadedFile) {
       send('error', { message: 'No file provided' });
-      return res.end();
+      return end();
     }
 
     const content = await fs.readFile(uploadedFile.filepath, 'utf-8');
@@ -54,7 +39,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
     if (!data.version || !data.expenses || !data.users) {
       send('error', { message: 'Invalid SplitPro backup file' });
-      return res.end();
+      return end();
     }
 
     const importFn = mode === 'restore' ? restoreSplitProData : importSplitProData;
@@ -71,6 +56,6 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
     if (uploadedFile) {
       await fs.unlink(uploadedFile.filepath).catch(() => null);
     }
-    res.end();
+    end();
   }
 }
